requirements.txt
pandas==2.0.0
kaggle~=1.5.12
prefect==2.10.3
tqdm~=4.64.1
torch~=1.13.1
matplotlib~=3.7.0
datasets~=2.9.0
transformers~=4.26.1
omegaconf~=2.3.0
nltk~=3.8.1
torchtext~=0.14.1
Pillow~=9.4.0
torchvision~=0.14.1
setuptools~=67.6.1
hydra-core~=1.3.2
pycocotools~=2.0.6
importlib-resources
requests~=2.28.2
pytest~=7.3.1